getAttitude remake and tests
